// src/types/ade.ts
export type AdeOption = {
    type: string;
    params: [];
    dtstamp: string;
    start: Date;
    datetype: string;
    end: Date;
    summary: string;
    location: string;
    description: string;
    uid: string;
    created: Date;
    lastmodified: Date;
    sequence: string;
    method: string;
};